Use LXD rather than docker for tests
We previously used docker to emulate a VM for tests, and used the host's
docker rather than running docker-in-docker. This was fast and worked
in GHA and in developer's environments.

However, now we are looking to test running docker services inside our
backend VMs, with file system mounting and port mapping, using sibling
docker was no longer viable.

Rather than mess with unsupported docker-in-docker, we instead switch to
LXD to provide lightweight VM-like environments for our tests. This more
closely models our actual deployment environments, which is good, and
allows us to properly test our docker deployments.

However, it has two main drawbacks: a) pulling images is really slow
because we're no longer using the hosts cache, and b) LXD is not as easy
to set up for devs or in GHA.

We've addressed a) by removing the pre-pull step from the jobrunner
install scripts. This works because jobrunner will start pulling these
automatically soon
(https://github.com/opensafely-core/job-runner/pull/250). Along with
some manual LXD image building and caching, test speed locally is
comparable to docker (although building the image takes a bit longer).

For b), its simple enough in GHA, and we'll just have to suck it up in
dev. Devs can either install lxd locally (which is easy on ubuntu, less
so on Macs), or use the GHA environment to run tests.