make release-hatch run over TLS by default.
The release-hatch API is used over CORS from a browser in a HTTPS
context, and as such, we need it to have TLS enabled or we get all kinds
of issues.

This PR changes the set up for release-hatch to always run over TLS with
a self-signed cert by default.

We anticipate this cert being replaced with a real one of some kind or
other in situ, but having self signed by default allows us to test our
TLS config locally.