Add timers service
This allows for arbitrary regular scheduled jobs to be run.

The first one is vmp-mapping, which only runs on the TPP backend. It
consists of 2 systemd unit files and script to run.

We test that this is installed and enabled correctly, and than when run,
errors with the error we expect for running in test env