Move jobrunner to using a docker image with autodeployment.
This is patterned on how we run airlock:
 - uses docker compose to run and manage restarts and logs
 - a systemd timer/service to regularly pull down new images

We get to remove the code/lib dirs, and some of the awkward scripting.

Jobrunner has a lot of management commands, but those are mostly
abstracted behind just, and I have added any that were missing to the
justfile.

Deploying this will require some post install manuall cleanup of
code/lib dirs.

It was difficult to break up into small commits, at its a big refactor
that touches a lot of related things.