Temporarily remove `readonly` from BACKEND var
This caused us problems because a later script attempts to set `backend`
again (to the same value) causing the whole thing to error out.

The proper solution here is to refactor things so it's clear exactly
which parts of the system are responsible for setting which environment
variables (see #257). But we did the quick fix directly on the server,
and this change makes the bootstrap script reflect the current state of
the machine which seems the safest way to park things for now.

Associated Slack threads:
https://bennettoxford.slack.com/archives/C069YDR4NCA/p1733144179614949
https://bennettoxford.slack.com/archives/C02GL3A9THD/p1731000367418999