Adjust default API poll time.
The default for API poll was 15s originally, but has been overridden to
30s in TPP for a while. This changes it to 60s.

Given the run times of jobs measuring hours, communicating a state
change within 15s seems unecessarily highly tuned.

Also, when we have large numbers of jobs, the API payload is large and
it can add a fair bit of load to job-server[1], when most of the calls,
there are actually zero state changes to process. However, the API poll
in job-server is self-throttling here so it's maybe not a big impact.

Still, the test backend is hitting job-server 4 times minute, which
seems a bit excessive.

[1] Example - in the two busy periods, latency jobs from ~5s 5o ~20s
https://ui.honeycomb.io/bennett-institute-for-applied-data-science/environments/production/datasets/job-server/result/cFfSbbPXsmE?hideCompare=