Replace agent-controller test with agent-only test
This is obviously a much more anaemic test that what went before but we
now can't easily do a full end-to-end test because that requires a
component which this repo no longer manages.

We might want to consider doing something more involved here like
spinning up a controller container purely for test purposes, but I don't
want to block on that for now.

I initially tried just starting the agent and ignoring the controller
entirely, but the agent won't stay running if it can't talk to the
controller and so to get the tests to pass it needs to have _something_
to connect to. Faking it with `http.server` seemed to like the least
effort way to achieve this.