Upgrade to latest honeycomb metrics build/config
https://github.com/honeycombio/opentelemetry-collector-configs

This adds some new filesystem usage things, will look at what they show
when we deploy, moves to newer config.

I also stopped pulling the binary from github at runtime, and instead
check it in to this repo, like we do with other binaries. This means we
can disable the proxy rule allowing this, and standardises on one way to
ship binaries into the backends.