Set the agent and airlock service names separately.
This is not ideal, as we need to read environment variables in the
docker compose file, which only supports 1 file, so our multiple env
files don't work in it.

Workaround is to write the service name into a .env at install time. Not
ideal, I think longer term we need to rework how we use docker compose,
but will do for now.