Bump otel limits
The default client limits for attributes per span is 128, and size of
attribute value is 1024.

If you hit them, the additional attributes are dropped, and the values
are truncated.

Honeycomb supports much higher values, so set them globally to ensure we
do not silently lose data.

The large value length is particularly useful for ehrql's preformatted
SQL queries.