Run airlock jobs daily
Run the airlock jobs daily. Note this is slightly confusing, because
the jobs are in the "weekly" group, however, this is just a grouping
that we can run on any schedule we like. In this case, we expect these
jobs to need to do something once a week, but we run them daily so
that we can do that thing whenever it's needed, and don't have to
wait until one specific time per week.